fix: pagem割のリファクタリング
